Update patch set 1
Patch Set 1:

I did some tests to have Requires only for centos package, and I did not find a good solution.

It easy to add Requires for a subpackage, but we want to do the opposite. It should be possible to manage that by doing:

%package centos
Requires: centos-queens
%package rhel

But that means we will have 2 packages at the end of the build process
sf-release-centos and sf-release-rhel

I will check how to manage that's package installation in sf-config instead and remove Requires statement from sf-release

Patch-set: 1